+++ spec/spec_helper.rb
@@ -1,15 +1,27 @@
+# frozen_string_literal: true
+
 require 'chefspec'
 require 'chefspec/berkshelf'
 
-def not_windows?
-  (RUBY_PLATFORM =~ /mswin|mingw|windows/).nil?
+Dir[File.join(__dir__, '..', 'libraries', '*.rb')].sort.each { |f| require File.expand_path(f) }
+
+if ENV['CHEFSPEC_NO_ZERO_SERVER']
+  module ChefSpec
+    class ZeroServer
+      class << self
+        def setup!; end
+
+        def reset!; end
+
+        def teardown!; end
+      end
+    end
+  end
 end
 
-# Require all our libraries
-Dir['libraries/*.rb'].each { |f| require File.expand_path(f) }
-
 RSpec.configure do |config|
-  config.color = true               # Use color in STDOUT
-  config.formatter = :documentation # Use the specified formatter
-  config.log_level = :error         # Avoid deprecation notice SPAM
+  config.color = true
+  config.formatter = :documentation
+  config.log_level = :error
+  config.server_runner_port = 0
 end